requirements.txt
# Pinned, Python 3.12 compatible (Ubuntu 24.04).
# If you need CUDA wheels, set TORCH_INDEX_URL when installing (see scripts/bootstrap_venv.sh).
numpy==1.26.4
pyyaml==6.0.2
tqdm==4.66.6
pillow==10.4.0
scipy==1.12.0
pandas==2.2.3

# Core ML
torch==2.5.1
torchvision==0.20.1

# COCO evaluation / utilities
pycocotools==2.0.8
opencv-python==4.10.0.84

